+++ src/main/java/janggi/domain/piece/movement/strategy/FriendlyPalaceSingleStepStrategy.java
@@ -1,4 +1,4 @@
-package janggi.domain.piece.strategy;
+package janggi.domain.piece.movement.strategy;
 
 import janggi.domain.board.Palace;
 import janggi.domain.board.Position;
