+++ src/test/java/janggi/domain/piece/strategy/HorseStrategyTest.java
@@ -3,6 +3,8 @@ package janggi.domain.piece.strategy;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import janggi.domain.board.Position;
+import janggi.domain.piece.movement.strategy.HorseStrategy;
+import janggi.domain.piece.movement.strategy.MoveStrategy;
 import java.util.List;
 import java.util.stream.Stream;
 import org.assertj.core.api.SoftAssertions;
